requirements/base.txt
Django==5.2.7
asgiref==3.10.0
django-environ==0.12.0
django-cors-headers==4.9.0
django-filter==25.2
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
pillow==11.3.0
whitenoise==6.11.0

# TMDB Integration
tmdbv3api==1.9.0
requests==2.32.3

# ML Recommendations - Python 3.13 compatible
scikit-learn==1.6.1
pandas==2.2.3
scipy==1.15.1
numpy==2.2.6
